Desktop/pitchperfect/cover_letter_backend/requirements.txt
python-dotenv
streamlit
PyPDF2
requests
fpdf2